Make mm algo configurable and modular
This merely moves the existing algorithm that ignore the OS into its own
module. But open the gates for smarter algos.
I don't see us using many algos, but it's useful for testing mm
interaction to keep this algo ignoring OS around, when the focus is not
on the matching logic.
