+++ lib/teiserver/matchmaking/queue_server.ex
@@ -17,6 +17,7 @@ defmodule Teiserver.Matchmaking.QueueServer do
   alias Teiserver.Player
   alias Teiserver.Helpers.MonitorCollection, as: MC
   alias Teiserver.Matchmaking.Member
+  alias Teiserver.Matchmaking.Algos
 
   @type id :: String.t()
 
@@ -31,6 +32,11 @@ defmodule Teiserver.Matchmaking.QueueServer do
         }
 
   @typedoc """
+  what algorithm this queue should use to put players/party together?
+  """
+  @type algo :: :ignore_os
+
+  @typedoc """
   immutable specification of the queue
   """
   @type queue :: %{
@@ -38,6 +44,7 @@ defmodule Teiserver.Matchmaking.QueueServer do
           team_size: pos_integer(),
           team_count: pos_integer(),
           ranked: boolean(),
+          algo: algo(),
           engines: [%{version: String.t()}],
           games: [%{spring_game: String.t()}],
           maps: [Teiserver.Asset.Map.t()]
@@ -84,7 +91,8 @@ defmodule Teiserver.Matchmaking.QueueServer do
           optional(:games) => [%{spring_game: String.t()}],
           optional(:maps) => [Teiserver.Asset.Map.t()],
           optional(:settings) => settings(),
-          optional(:members) => [Member.t()]
+          optional(:members) => [Member.t()],
+          optional(:algo) => algo()
         }) :: state()
   def init_state(attrs) do
     %{
@@ -93,6 +101,7 @@ defmodule Teiserver.Matchmaking.QueueServer do
         name: attrs.name,
         team_size: attrs.team_size,
         team_count: attrs.team_count,
+        algo: Map.get(attrs, :algo, :ignore_os),
         ranked: true,
         engines: Map.get(attrs, :engines, []),
         games: Map.get(attrs, :games, []),
@@ -239,7 +248,7 @@ defmodule Teiserver.Matchmaking.QueueServer do
             rating: %{},
             # TODO: get and aggregate avoids somehow
             avoid: [],
-            joined_at: DateTime.utc_now(),
+            joined_at: DateTime.utc_now()
           }
 
           :timer.cancel(pending.tref)
@@ -535,42 +544,9 @@ defmodule Teiserver.Matchmaking.QueueServer do
   """
   @spec match_members(state()) :: :no_match | {:match, [[[Member.t()]]]}
   def match_members(state) do
-    case greedy_match(state.queue.team_size, state.queue.team_count, state.members, [], []) do
-      [] -> :no_match
-      matches -> {:match, matches}
-    end
-  end
-
-  defp greedy_match(team_size, team_count, members, current_team, matched) do
-    # tachyon_mvp: this is a temporary algorithm
-    # it only looks at the number of players to fill a team
-    case members do
-      [] ->
-        Enum.chunk_every(matched, team_count, team_count, :discard)
-
-      members ->
-        current_size =
-          current_team |> Enum.map(fn member -> Enum.count(member.player_ids) end) |> Enum.sum()
-
-        case Enum.split_while(members, fn m ->
-               Enum.count(m.player_ids) + current_size > team_size
-             end) do
-          # current team cannot be completed, discard it
-          {_, []} ->
-            greedy_match(team_size, team_count, members, [], matched)
-
-          {too_big, [member | rest]} ->
-            to_add = Enum.count(member.player_ids)
-            rest = too_big ++ rest
-
-            cond do
-              current_size + to_add < team_size ->
-                greedy_match(team_size, team_count, rest, [member | current_team], matched)
-
-              current_size + to_add == team_size ->
-                greedy_match(team_size, team_count, rest, [], [[member | current_team] | matched])
-            end
-        end
+    case state.queue.algo do
+      :ignore_os ->
+        Algos.ignore_os(state.queue.team_size, state.queue.team_count, state.members)
     end
   end
